add step-9
this step-9 expand our commands from the step before
to a gap between multiple eco systems: Nodejs, FSharp and Rust.
we added model.fsx that will define some simple types to shared across
our NodeJS server and our Rust CLI tool we added commands to consume the
server with curl and pipe it to our rust cli which in turn will parse
the json with the common types transpiled from model.fsx

we will can git ignore fable_module for the JS eco system becuase it
does not need to create a "nix flake package", we wont be able to ignore
for now the fable_modules in the rust project due to the fact they are
"counted" as part of the source code to build the rust project.
there are way to solve it but for sake of simplicity I am going to leave
it for now.